Revert "Add a boatload of new tests."
This reverts commit 3c084e709ae5e94646240505f55a1861b138e038.
