+++ api/tests/python/end_points/test_auto_analyse.py
@@ -119,7 +119,7 @@ def test_new_bc_cr_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY INC.',  # OR [INCORPORATED, LTD]
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'CR',
         'request_type': 'NEW'
@@ -141,7 +141,7 @@ def test_new_bc_ul_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY ULC.',  # OR [UNLIMITED LIABILITY COMPANY]
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'UL',
         'request_type': 'NEW'
@@ -167,7 +167,7 @@ def test_new_bc_cp_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY COOP',  # []
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'CP',
         'request_type': 'NEW'
@@ -189,7 +189,7 @@ def test_new_bc_bc_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY INC.',  # OR [INCORPORATED, LTD]
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'BC',
         'request_type': 'NEW'
@@ -211,7 +211,7 @@ def test_new_bc_cc_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY INC.',  # OR [INCORPORATED, LTD] ALSO REQ *CCC* INC OR COMMUNITY CONTRIBUTION COMPANY
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'CC',
         'request_type': 'NEW'
@@ -233,7 +233,7 @@ def test_new_bc_fr_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY',
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'FR',
         'request_type': 'NEW'
@@ -255,7 +255,7 @@ def test_new_bc_dba_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY',
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'DBA',
         'request_type': 'NEW'
@@ -277,7 +277,7 @@ def test_new_bc_gp_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY',
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'GP',
         'request_type': 'NEW'
@@ -299,7 +299,7 @@ def test_new_bc_lp_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY LP.',
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'LP',
         'request_type': 'NEW'
@@ -321,7 +321,7 @@ def test_new_bc_ll_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY LLP.',
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'BC',
         'entity_type': 'LL',
         'request_type': 'NEW'
@@ -343,7 +343,7 @@ def test_new_xpro_xcr_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY INC.',  # OR LTD.
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'CA',
         'entity_type': 'XCR',
         'request_type': 'NEW'
@@ -365,7 +365,7 @@ def test_new_xpro_xul_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY ULC.',
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'CA',
         'entity_type': 'XUL',
         'request_type': 'NEW'
@@ -387,7 +387,7 @@ def test_new_xpro_xcp_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY COOP',
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'CA',
         'entity_type': 'XCP',
         'request_type': 'NEW'
@@ -409,7 +409,7 @@ def test_new_xpro_xlc_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY LLC.',  # CHECK FOR DESIGNATION
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'CA',
         'entity_type': 'XLC',
         'request_type': 'NEW'
@@ -431,7 +431,7 @@ def test_new_xpro_xlp_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY LP.',  # CHECK FOR DESIGNATION
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'CA',
         'entity_type': 'XLC',
         'request_type': 'NEW'
@@ -453,7 +453,7 @@ def test_new_xpro_xll_valid_response(client, jwt, app):
     headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
 
     test_params = {
-        'name': 'BOB\'S CARPENTRY LLP.',
+        'name': 'BOB\'S CARPENTRY INC.',
         'location': 'CA',
         'entity_type': 'XLL',
         'request_type': 'NEW'
@@ -469,3 +469,229 @@ def test_new_xpro_xll_valid_response(client, jwt, app):
         assert_issues_count_is(0, payload.issues)
 
 # IN THIS SECTION TEST VARIOUS ERROR RESPONSES
+
+
+# @pytest.mark.xfail(raises=ValueError)
+def test_add_distinctive_word_request_response(client, jwt, app):
+    # create JWT & setup header with a Bearer Token using the JWT
+    token = jwt.create_jwt(claims, token_header)
+    headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
+
+    test_params = {
+        'name': 'BOB\'S CARPENTRY INC.',
+        'location': 'BC',
+        'entity_type': 'FR',
+        'request_type': 'NEW'
+    }
+
+    query = '&'.join("{!s}={}".format(k, quote_plus(v)) for (k, v) in test_params.items())
+    path = ENDPOINT_PATH + '?' + query
+    print('\n' + 'request: ' + path + '\n')
+    response = client.get(path, headers=headers)
+    payload = jsonpickle.decode(response.data)
+    print("Assert that the payload contains issues")
+    if isinstance(payload.issues, list):
+        assert_issues_count_is_gt(0, payload.issues)
+
+        for issue in payload.issues:
+            # Make sure only Well Formed name issues are being returned
+            assert_issue_type_is_one_of([
+                AnalysisResultCodes.ADD_DISTINCTIVE_WORD,
+                AnalysisResultCodes.ADD_DESCRIPTIVE_WORD,
+                AnalysisResultCodes.TOO_MANY_WORDS
+            ], issue)
+
+        assert_has_issue_type(AnalysisResultCodes.ADD_DISTINCTIVE_WORD, payload.issues)
+
+
+# @pytest.mark.xfail(raises=ValueError)
+def test_add_descriptive_word_request_response(client, jwt, app):
+    # create JWT & setup header with a Bearer Token using the JWT
+    token = jwt.create_jwt(claims, token_header)
+    headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
+
+    test_params = {
+        'name': 'BOB\'S CARPENTRY INC.',
+        'location': 'BC',
+        'entity_type': 'FR',
+        'request_type': 'NEW'
+    }
+
+    query = '&'.join("{!s}={}".format(k, quote_plus(v)) for (k, v) in test_params.items())
+    path = ENDPOINT_PATH + '?' + query
+    print('\n' + 'request: ' + path + '\n')
+    response = client.get(path, headers=headers)
+    payload = jsonpickle.decode(response.data)
+    print("Assert that the payload contains issues")
+    if isinstance(payload.issues, list):
+        assert_issues_count_is_gt(0, payload.issues)
+
+        for issue in payload.issues:
+            # Make sure only Well Formed name issues are being returned
+            assert_issue_type_is_one_of([
+                AnalysisResultCodes.ADD_DISTINCTIVE_WORD,
+                AnalysisResultCodes.ADD_DESCRIPTIVE_WORD,
+                AnalysisResultCodes.TOO_MANY_WORDS
+            ], issue)
+
+        assert_has_issue_type(AnalysisResultCodes.ADD_DESCRIPTIVE_WORD, payload.issues)
+
+
+# @pytest.mark.xfail(raises=ValueError)
+def test_too_many_words_request_response(client, jwt, app):
+    # create JWT & setup header with a Bearer Token using the JWT
+    token = jwt.create_jwt(claims, token_header)
+    headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
+
+    test_params = {
+        'name': 'BOB\'S CARPENTRY AND HOME RENOVATIONS INC.',
+        'location': 'BC',
+        'entity_type': 'FR',
+        'request_type': 'NEW'
+    }
+
+    query = '&'.join("{!s}={}".format(k, quote_plus(v)) for (k, v) in test_params.items())
+    path = ENDPOINT_PATH + '?' + query
+    print('\n' + 'request: ' + path + '\n')
+    response = client.get(path, headers=headers)
+    payload = jsonpickle.decode(response.data)
+    print("Assert that the payload contains issues")
+    if isinstance(payload.issues, list):
+        assert_issues_count_is_gt(0, payload.issues)
+
+        for issue in payload.issues:
+            # Make sure only Well Formed name issues are being returned
+            assert_issue_type_is_one_of([
+                AnalysisResultCodes.ADD_DISTINCTIVE_WORD,
+                AnalysisResultCodes.ADD_DESCRIPTIVE_WORD,
+                AnalysisResultCodes.TOO_MANY_WORDS
+            ], issue)
+
+        assert_has_issue_type(AnalysisResultCodes.TOO_MANY_WORDS, payload.issues)
+
+
+# @pytest.mark.xfail(raises=ValueError)
+def test_contains_words_to_avoid_request_response(client, jwt, app):
+    # create JWT & setup header with a Bearer Token using the JWT
+    token = jwt.create_jwt(claims, token_header)
+    headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
+
+    test_params = {
+        'name': 'BOB\'S BULLSHIT CARPENTRY INC.',
+        'location': 'BC',
+        'entity_type': 'CR',
+        'request_type': 'NEW'
+    }
+
+    query = '&'.join("{!s}={}".format(k, quote_plus(v)) for (k, v) in test_params.items())
+    path = ENDPOINT_PATH + '?' + query
+    print('\n' + 'request: ' + path + '\n')
+    response = client.get(path, headers=headers)
+    payload = jsonpickle.decode(response.data)
+    print("Assert that the payload contains issues")
+    if isinstance(payload.issues, list):
+        assert_issues_count_is_gt(0, payload.issues)
+        assert_has_issue_type(AnalysisResultCodes.WORD_TO_AVOID, payload.issues)
+
+
+# @pytest.mark.xfail(raises=ValueError)
+def test_designation_mismatch_request_response(client, jwt, app):
+    # create JWT & setup header with a Bearer Token using the JWT
+    token = jwt.create_jwt(claims, token_header)
+    headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
+
+    test_params = {
+        'name': 'BOB\'S CARPENTRY INC.',
+        'location': 'BC',
+        'entity_type': 'CR',
+        'request_type': 'NEW'
+    }
+
+    query = '&'.join("{!s}={}".format(k, quote_plus(v)) for (k, v) in test_params.items())
+    path = ENDPOINT_PATH + '?' + query
+    print('\n' + 'request: ' + path + '\n')
+    response = client.get(path, headers=headers)
+    payload = jsonpickle.decode(response.data)
+    print("Assert that the payload contains issues")
+    if isinstance(payload.issues, list):
+        assert_issues_count_is_gt(0, payload.issues)
+        assert_has_issue_type(AnalysisResultCodes.DESIGNATION_MISMATCH, payload.issues)
+
+    test_params = {
+        'name': 'MOUNTAIN NEW FOOD GROWERS INC.',
+        'location': 'CA',
+        'entity_type': 'XLC',
+        'request_type': 'NEW'
+    }
+
+# @pytest.mark.xfail(raises=ValueError)
+def test_name_requires_consent_request_response(client, jwt, app):
+    # create JWT & setup header with a Bearer Token using the JWT
+    token = jwt.create_jwt(claims, token_header)
+    headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
+
+    test_params = {
+        'name': 'BOB\'S CARPENTRY INC.',
+        'location': 'BC',
+        'entity_type': 'CR',
+        'request_type': 'NEW'
+    }
+
+    query = '&'.join("{!s}={}".format(k, quote_plus(v)) for (k, v) in test_params.items())
+    path = ENDPOINT_PATH + '?' + query
+    print('\n' + 'request: ' + path + '\n')
+    response = client.get(path, headers=headers)
+    payload = jsonpickle.decode(response.data)
+    print("Assert that the payload contains issues")
+    if isinstance(payload.issues, list):
+        assert_issues_count_is_gt(0, payload.issues)
+        assert_has_issue_type(AnalysisResultCodes.NAME_REQUIRES_CONSENT, payload.issues)
+
+# IN THIS SECTION TEST VARIOUS ERROR RESPONSES
+
+# @pytest.mark.xfail(raises=ValueError)
+def test_contains_unclassifiable_word_request_response(client, jwt, app):
+    # create JWT & setup header with a Bearer Token using the JWT
+    token = jwt.create_jwt(claims, token_header)
+    headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
+
+    test_params = {
+        'name': 'BOB\'S CARPENTRY INC.',
+        'location': 'BC',
+        'entity_type': 'FR',
+        'request_type': 'NEW'
+    }
+
+    query = '&'.join("{!s}={}".format(k, quote_plus(v)) for (k, v) in test_params.items())
+    path = ENDPOINT_PATH + '?' + query
+    print('\n' + 'request: ' + path + '\n')
+    response = client.get(path, headers=headers)
+    payload = jsonpickle.decode(response.data)
+    print("Assert that the payload contains issues")
+    if isinstance(payload.issues, list):
+        assert_issues_count_is_gt(0, payload.issues)
+        assert_has_issue_type(AnalysisResultCodes.CONTAINS_UNCLASSIFIABLE_WORD, payload.issues)
+
+
+# @pytest.mark.xfail(raises=ValueError)
+def test_corporate_name_conflict_request_response(client, jwt, app):
+    # create JWT & setup header with a Bearer Token using the JWT
+    token = jwt.create_jwt(claims, token_header)
+    headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
+
+    test_params = {
+        'name': 'BOB\'S CARPENTRY INC.',
+        'location': 'BC',
+        'entity_type': 'FR',
+        'request_type': 'NEW'
+    }
+
+    query = '&'.join("{!s}={}".format(k, quote_plus(v)) for (k, v) in test_params.items())
+    path = ENDPOINT_PATH + '?' + query
+    print('\n' + 'request: ' + path + '\n')
+    response = client.get(path, headers=headers)
+    payload = jsonpickle.decode(response.data)
+    print("Assert that the payload contains issues")
+    if isinstance(payload.issues, list):
+        assert_issues_count_is_gt(0, payload.issues)
+        assert_has_issue_type(AnalysisResultCodes.CORPORATE_CONFLICT, payload.issues)
